Downloading checkpoints from AML if not found on disk (#614)
Workaround for a temporary issue with low-priority preemption: checkpoint files are not available on disk upon job restart. Trying to download from AML.
