+++ Tests/ML/test_download_upload.py
@@ -12,7 +12,8 @@ from InnerEye.Azure.azure_config import AzureConfig
 from InnerEye.Common.common_util import OTHER_RUNS_SUBDIR_NAME, logging_to_stdout
 from InnerEye.Common.output_directories import OutputFolderForTests
 from InnerEye.ML.model_config_base import ModelConfigBase
-from InnerEye.ML.utils.run_recovery import RunRecovery
+from InnerEye.ML.utils.checkpoint_handling import download_all_checkpoints_from_run, \
+    download_best_checkpoints_from_child_runs
 from Tests.AfterTraining.test_after_training import FALLBACK_ENSEMBLE_RUN, FALLBACK_SINGLE_RUN, get_most_recent_run
 from Tests.ML.util import get_default_azure_config
 
@@ -43,7 +44,7 @@ def test_download_recovery_single_run(test_output_dirs: OutputFolderForTests,
     config = ModelConfigBase(should_validate=False)
     config.set_output_to(output_dir)
     run = get_most_recent_run(fallback_run_id_for_local_execution=FALLBACK_SINGLE_RUN)
-    run_recovery = RunRecovery.download_all_checkpoints_from_run(config, run)
+    run_recovery = download_all_checkpoints_from_run(config, run)
 
     # This fails if there is no recovery checkpoint
     check_single_checkpoint(run_recovery.get_recovery_checkpoint_paths())
@@ -58,7 +59,7 @@ def test_download_best_checkpoints_ensemble_run(test_output_dirs: OutputFolderFo
     config.set_output_to(output_dir)
 
     run = get_most_recent_run(fallback_run_id_for_local_execution=FALLBACK_ENSEMBLE_RUN)
-    run_recovery = RunRecovery.download_best_checkpoints_from_child_runs(config, run)
+    run_recovery = download_best_checkpoints_from_child_runs(config, run)
     other_runs_folder = config.checkpoint_folder / OTHER_RUNS_SUBDIR_NAME
     assert other_runs_folder.is_dir()
     for child in ["0", "1"]:
